third_party/blink/web_tests/http/tests/websocket/extensions-expected.txt

Test WebSocket.extensions attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check the value before opening the connection:
PASS ws.extensions is ""
Check if the value is read only:
PASS ws.extensions = 'foo'; ws.extensions is ""
Check the value after the connection is established:
PASS ws.extensions is "permessage-deflate"
Check the value of the Sec-WebSocket-Extensions header sent to the server:
PASS extensions_requested is "permessage-deflate; client_max_window_bits"
PASS successfullyParsed is true

TEST COMPLETE
